requirements.txt
PySide6==6.10.0
pandas
requests==2.31.0
cryptography==41.0.7
python-dotenv==1.0.0
reportlab==4.0.7
pycryptodome==3.19.0
XlsxWriter
matplotlib
matplotlib-venn
psutil
selenium
webdriver_manager
playwright
# Cloud admin dashboard
flask>=2.0.0
supabase>=2.0.0
gunicorn>=21.0.0